requirements.txt
asgiref==3.2.10
Django==3.0.8
django-currentuser==0.5.0
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
